refactor(ui): split sandboxes.$id.tsx god object into sandbox-detail/ tab modules
