+++ scripts/.god-object-baseline.txt
@@ -1,7 +1,6 @@
 # God-object baseline (LOC ceilings). Files here may only SHRINK.
 # Regenerate after a real split: scripts/check-file-sizes.sh --update
 # Doctrine: CLAUDE.md 'Module Size Discipline' — split by responsibility toward <600 LOC.
-1281 ui/src/routes/sandboxes.$id.tsx
 1106 ui/src/routes/instances.$id.tsx
 1095 ai-agent-sandbox-blueprint-lib/src/workflows.rs
 1065 sandbox-runtime/src/session_auth.rs
@@ -9,6 +8,7 @@
 922 ui/src/routes/workflows._index.tsx
 838 sandbox-runtime/src/reaper.rs
 744 sandbox-runtime/src/scoped_session_auth.rs
+738 ui/src/routes/sandboxes.$id.tsx
 721 ui/src/components/shared/SessionSidebar.tsx
 701 ui/src/lib/stores/chatSessions.ts
 682 ui/src/components/shared/InfrastructureModal.tsx
